+++ extensions/twitter/src/mytweets.tsx
@@ -1,5 +1,8 @@
+import { withXAuth } from "./v2/lib/with_x_auth";
 import { MyTweetListV2 } from "./v2/components/mytweets";
 
-export default function MyTweetRoot() {
+function MyTweetRoot() {
   return <MyTweetListV2 />;
 }
+
+export default withXAuth(MyTweetRoot);